Issue #4: Initial version of service
The service listens for HTTP requests and calls Jenkins.

It currently does not interpret the request, but treats it as a branch creation event.
It uses a hard-coded URL to Jenkins.

There is a tiny acceptance test which requires that the service has been started manually.
The test uses curl to send a HTTP request to the service.