Issue #4: Trigger Jenkins parameterized build with branch name
Update seed job config and DSL script to take a BRANCH parameter and use it in the
names of the generated pipeline jobs.

Update the FlowAbstractionLayer to apply the branch name, but in a sanitized version:
* Branch feature, e.g. 'feature/', is removed
* Dashes are converted to underscores
This is done to avoid warnings and errors in Jenkins.

Updated the acceptance test to poll for the new job names.