Issue #4: External configuration file in YAML format
Change the configuration file format from Java properties to YAML.
Move the file to config/ folder and pass it as argument when running the service.

The file is parsed in Java using the snakeyaml library.

The file still needs to be updated with a mapping from branch names to pipeline contents.